IDA: Support lookup of stack variables using `$ida("name")` (#2926)
Instead of only allowing to lookup global symbols, allow reference local variables too.

This is done by iterating the local stack frame struct in IDA and comparing the name. Depending on whether IDA created  an entry for the saved frame pointer in the stack frame, the stack variable offset is returned relative to the current stack or frame pointer.

This also fixes the error for invalid names in #2903

Based on https://gist.github.com/syndrowm/2968620